requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
pygame>=2.5.0
matplotlib>=3.8.0
pandas>=2.1.0
Pillow>=10.0.0